+++ setup.py
@@ -46,7 +46,7 @@ setup(
     extras_require={
         "glue": ["boto3", "mypy-boto3-glue"],
         "md": ["duckdb==1.0.0"],
-        "unity": ["unitycatalog==0.1.1"],
+        "unity": ["unitycatalog==0.1.1", "deltalake==0.18.2", "pyarrow==17.0.0"],
         "delta": ["deltalake==0.18.2", "pyarrow==17.0.0"],
     },
     classifiers=[
