+++ tests/functional/plugins/test_delta_write.py
@@ -1,6 +1,5 @@
 import tempfile
 from pathlib import Path
-from unittest.mock import patch
 
 import pytest
 from dbt.tests.util import (
